+++ components/datetime/src/pattern/error.rs
@@ -3,15 +3,14 @@ use crate::fields;
 #[derive(Debug)]
 pub enum Error {
     Unknown,
-    FieldTooLong,
+    FieldTooLong(fields::FieldSymbol),
     UnknownSubstitution(u8),
 }
 
 impl From<fields::Error> for Error {
     fn from(input: fields::Error) -> Self {
         match input {
-            fields::Error::Unknown => Self::Unknown,
-            fields::Error::TooLong => Self::FieldTooLong,
+            fields::Error::TooLong(symbol) => Self::FieldTooLong(symbol),
         }
     }
 }
